0-print_list.c

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -1,5 +1,6 @@
 #include <stdio.h>
 #include "lists.h"
+#include <stdlib.h>
 
 /**
  * print_list - Prints all the elements of a list_t list
@@ -10,7 +11,7 @@
 
 size_t print_list(const list_t *h)
 {
-	int count = 0;
+	size_t count = 0;
 
 	while (h != NULL)
 	{
